web/requirements-updated.txt
# Updated dependencies for Python 3.12 compatibility and security
# Last updated: November 2025

# Web Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Azure SDK
azure-servicebus==7.12.0
azure-identity==1.15.0

# Email
sendgrid==6.11.0
python-http-client==3.3.7

# Utilities
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.1.0

# Security updates
cryptography==41.0.7
PyJWT==2.8.0
certifi==2023.11.17
